More progress on interactive support.
Since we have ranges, and since we don't have a good way to push errors
to Flycheck for display, we probably could use our own parsing. If
https://github.com/flycheck/flycheck/issues/527 gets implemented,
though, it would be nice to revisit this. There's no hurry though, since
it's not clear how well the interactive mode would blend with Flycheck's
model.