Add a new Flycheck checker and enable Flycheck by default
The original Flycheck checker invokes a fully separate F* instance and
re-verifies the entire file every time.  The new checker is different: it works
alongside the interactive PG-style verification.  Using a new interactive-mode
command, it sends just a few lines to the F* subprocess and gets results of
light syntax- and type-checking in return.

This uses :predicate instead of :enabled to make it easier to switch between
multiple F* versions.