Ensure that the feature list is immediately available
Loading the feature list early lets us run feature checks in the first push
without having to add complex synchronization (see next commit for a use case).

Note that we already pay the price of one F* initialization anyway when running
fstar --version (we'll be able to remove that version check when the old
protocol is officially deprecated — in a few years), so blocking until we get
the feature list isn't too bad.