fstar-mode.el: do not pass --smt
Newer F* versions support many coexisting Z3 backends (we are mostly
using 4.8.5 and 4.13.3). Passing --smt makes F* pick a single given
executable for Z3, and prevents the version switching from working.
Remove it.

I think this was added to not need a Z3 in the path. If using a binary
package, F* will actually pick up z3 from an internal path under
/lib/fstar, so no Z3 need be in the PATH. For developers, z3 should be
found in the PATH with a name like `z3-4.13.3`.