init-compatibility-layer: fix F* version check
A valid F* version number must start with a number; on recent emacsen
(e.g. 29.4), `version<=` (in fact, `version-to-list`) will fail if
not. So, we fix the regexp accordingly.

In the case the F* version string does not parse properly, we also
assume the version number globally, instead of setting it to
"unknown". This way, we guarantee that `fstar--vernum` will always be
a valid version number.